--- a/src/cmd/distpack/pack.go
+++ b/src/cmd/distpack/pack.go
@@ -271,7 +271,7 @@ func readVERSION(goroot string) (version string, t time.Time) {
 		log.Fatal(err)
 	}
 	version, rest, _ := strings.Cut(string(data), "\n")
-	for _, line := range strings.Split(rest, "\n") {
+	for line := range strings.SplitSeq(rest, "\n") {
 		f := strings.Fields(line)
 		if len(f) == 0 {
 			continue