--- a/src/cmd/go/internal/auth/gitauth_test.go
+++ b/src/cmd/go/internal/auth/gitauth_test.go
@@ -5,6 +5,7 @@
 package auth
 
 import (
+	"strings"
 	"testing"
 )
 
@@ -80,5 +81,79 @@ password:secr3t
 		if password != tc.wantPassword {
 			t.Errorf("parseGitAuth(%s):\nhave %q\nwant %q", tc.gitauth, password, tc.wantPassword)
 		}
+	}
+}
+
+func BenchmarkParseGitAuth(b *testing.B) {
+	// Define different test scenarios to benchmark
+	testCases := []struct {
+		name string
+		data []byte
+	}{{
+		// Standard scenario with all basic fields present
+		name: "standard",
+		data: []byte(`
+protocol=https
+host=example.com
+username=bob
+password=secr3t
+`),
+	}, {
+		// Scenario with URL field included
+		name: "with_url",
+		data: []byte(`
+protocol=https
+host=example.com
+username=bob
+password=secr3t
+url=https://example.com/repo
+`),
+	}, {
+		// Minimal scenario with only required fields
+		name: "minimal",
+		data: []byte(`
+protocol=https
+host=example.com
+`),
+	}, {
+		// Complex scenario with longer values and extra fields
+		name: "complex",
+		data: func() []byte {
+			var builder strings.Builder
+			builder.WriteString("protocol=https\n")
+			builder.WriteString("host=example.com\n")
+			builder.WriteString("username=longusernamenamename\n")
+			builder.WriteString("password=longpasswordwithmanycharacters123456789\n")
+			builder.WriteString("url=https://example.com/very/long/path/to/repository\n")
+			builder.WriteString("extra1=value1\n")
+			builder.WriteString("extra2=value2\n")
+			return []byte(builder.String())
+		}(),
+	}, {
+		// Scenario with empty input
+		name: "empty",
+		data: []byte(``),
+	}, {
+		// Scenario with malformed input (using colon instead of equals)
+		name: "malformed",
+		data: []byte(`
+protocol:https
+host:example.com
+username:bob
+password:secr3t
+`),
+	}}
+
+	for _, tc := range testCases {
+		b.Run(tc.name, func(b *testing.B) {
+			b.ResetTimer()
+			for b.Loop() {
+				prefix, username, password := parseGitAuth(tc.data)
+
+				_ = prefix
+				_ = username
+				_ = password
+			}
+		})
 	}
 }