--- a/src/cmd/go/internal/doc/pkg.go
+++ b/src/cmd/go/internal/doc/pkg.go
@@ -920,7 +920,7 @@ func trimUnexportedFields(fields *ast.FieldList, isInterface bool) *ast.FieldLis
 
 			start := doc.List[0].Slash
 			doc.List = doc.List[:0]
-			for _, line := range strings.Split(text, "\n") {
+			for line := range strings.SplitSeq(text, "\n") {
 				prefix := "// "
 				if len(line) > 0 && line[0] == '\t' {
 					prefix = "//"