--- a/src/internal/buildcfg/exp.go
+++ b/src/internal/buildcfg/exp.go
@@ -113,7 +113,7 @@ func ParseGOEXPERIMENT(goos, goarch, goexp string) (*ExperimentFlags, error) {
 		}
 
 		// Parse names.
-		for _, f := range strings.Split(goexp, ",") {
+		for f := range strings.SplitSeq(goexp, ",") {
 			if f == "" {
 				continue
 			}