--- a/src/os/exec/lp_windows.go
+++ b/src/os/exec/lp_windows.go
@@ -123,7 +123,7 @@ func pathExt() []string {
 	var exts []string
 	x := os.Getenv(`PATHEXT`)
 	if x != "" {
-		for _, e := range strings.Split(strings.ToLower(x), `;`) {
+		for e := range strings.SplitSeq(strings.ToLower(x), `;`) {
 			if e == "" {
 				continue
 			}